Fixed accessing AppKit element from background thread.
When forking a new code snippet repo (Plug-ins -> Configure snippets
repo) when checking the console logs AppKit outputs the following entry:

NSAlert is being used from a background thread, which is not safe.
This is probably going to crash sometimes. Break on void
_NSAlertWarnUnsafeBackgroundThreadUsage() to debug.  This will be
logged only once.  This may break in the future.